requirements.txt
dnspython==2.7.0
psycopg2-binary==2.9.10
pymongo==4.12.0
python-dotenv==1.1.0
